+++ packages/nx-plugin/src/generators/app/generator.ts
@@ -109,7 +109,7 @@ export async function appGenerator(
   });
 
   const angularVersion = getInstalledPackageVersion(tree, '@angular/core');
-  const majorAngularVersion = major(coerce(angularVersion));
+  const majorAngularVersion = major(coerce(angularVersion!)!);
   addFiles(tree, normalizedOptions, majorAngularVersion);
   addDependenciesToPackageJson(
     tree,
@@ -122,11 +122,15 @@ export async function appGenerator(
     {},
   );
 
-  updateJson<{ dependencies: object }>(tree, '/package.json', (json) => {
-    json.dependencies['@angular/platform-server'] = `~${angularVersion}`;
+  updateJson<{ dependencies: Record<string, string> }>(
+    tree,
+    '/package.json',
+    (json) => {
+      json.dependencies['@angular/platform-server'] = `~${angularVersion}`;
 
-    return json;
-  });
+      return json;
+    },
+  );
 
   updateIndex(tree, normalizedOptions.analogAppName);