+++ packages/nx-plugin/src/generators/app/lib/initialize-analog-workspace.ts
@@ -40,7 +40,7 @@ export async function initializeAngularWorkspace(
     }
   }
 
-  if (belowMinimumSupportedAngularVersion(angularVersion)) {
+  if (belowMinimumSupportedAngularVersion(angularVersion!)) {
     throw new Error(
       stripIndents`Analog only supports an Angular version of 15 and higher`,
     );
@@ -82,7 +82,7 @@ const initWithNxNamespace = async (
     },
   );
 
-  return getInstalledPackageVersion(tree, '@angular/core', null, true);
+  return getInstalledPackageVersion(tree, '@angular/core', undefined, true);
 };
 
 const initWithNrwlNamespace = async (
@@ -117,5 +117,5 @@ const initWithNrwlNamespace = async (
     skipInstall: true,
     skipFormat: skipFormat,
   });
-  return getInstalledPackageVersion(tree, '@angular/core', null, true);
+  return getInstalledPackageVersion(tree, '@angular/core', undefined, true);
 };