+++ packages/nx-plugin/src/generators/init/lib/update-app-tsconfig.ts
@@ -13,7 +13,7 @@ export function updateAppTsConfig(
 ) {
   const projects = getProjects(tree);
 
-  const projectConfig = projects.get(schema.project);
+  const projectConfig = projects.get(schema.project)!;
 
   const tsconfigPath = joinPathFragments(
     projectConfig.root,