+++ packages/vite-plugin-angular-tools/src/builders/vite/vite-build.impl.ts
@@ -12,8 +12,8 @@ async function viteBuilder(
   context: BuilderContext,
 ): Promise<BuilderOutput> {
   const { createBuilder } = await Function('return import("vite")')();
-  const projectConfig = await context.getProjectMetadata(context.target);
-  const projectName = context.target.project;
+  const projectConfig = await context.getProjectMetadata(context.target!);
+  const projectName = context.target?.project;
   const configuration = context.target?.configuration || 'production';
   const buildTargetSpecifier = `::${configuration}`;
   const buildTarget = targetFromTargetString(